requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
psycopg2-binary==2.9.3
pika==1.3.0
openpyxl==3.0.10
openpyxl-dictreader==0.1.9